feat: enhance dashboard loading experience and integrate new spinner component
- Introduced a loading spinner component to improve user experience during data fetching.
- Updated Dashboard.vue to conditionally display the spinner while loading match history.
- Refactored match history fetching logic to utilize the new loading state management.
- Enhanced SummonerCard.vue to access the dark mode state from the app context.
- Added utility composable for managing loading states across components.
- Updated TypeScript definitions to include the new Spinner component and loading utility.

+++ src-tauri/src/lcu/analysis_data/service.rs
@@ -461,7 +461,7 @@ async fn fetch_all_players_match_stats(
                     player.display_name
                 );
 
-                match get_recent_matches_by_puuid(http_client, &summoner_info.puuid, 20, Some(queue_id)).await {
+                match get_recent_matches_by_puuid(http_client, &summoner_info.puuid, 20, Some(queue_id as i32)).await {
                     Ok(player_stats) => {
                         // 注意：get_recent_matches_by_puuid 已经返回完整的 PlayerMatchStats
                         // 包含了 traits, today_games 等所有增强字段
